requirements.txt
PyPDF2
python-docx
streamlit
pandas
openai
firebase-admin
